requirements.txt
speechrecognition
pyaudio
setuptools
pyttsx3
pocketsphinx
pyjokes
openai
python-dotenv
